projects/LANS-scubelic-asucur-lborovic-nmarjanovic/requirements.txt
# Core dependencies for llm-answer-watcher
# Generated from pyproject.toml
# For pip users - prefer using 'uv sync' if available

# Core runtime dependencies
pydantic[email]>=2.0
email-validator>=2.0.0
pyyaml>=6.0
httpx>=0.27.0
tenacity>=8.0
rapidfuzz>=3.0
jinja2>=3.1
typer>=0.12.0
rich>=13.0
fastapi>=0.110.0
uvicorn>=0.29.0
passlib[argon2]>=1.7.4
python-jose>=3.3.0
cryptography>=42.0.0

# Development dependencies (optional)
# Install with: pip install -r requirements.txt
# Or core only: pip install pydantic pyyaml httpx tenacity rapidfuzz jinja2 typer rich

# Uncomment below for development dependencies:
pytest>=8.0
pytest-httpx>=0.30.0
pytest-cov>=5.0
pytest-mock>=3.14.0
freezegun>=1.5.0
ruff>=0.4.0
